@@ -2,8 +2,8 @@ require_dependency "user_impersonate/application_controller"
 
 module UserImpersonate
   class ImpersonateController < ApplicationController
-    before_filter :authenticate_the_user, except: ["destroy"]
-    before_filter :current_user_must_be_staff!, except: ["destroy"]
+    before_action :authenticate_the_user, except: ["destroy"]
+    before_action :current_user_must_be_staff!, except: ["destroy"]
 
 
     helper_method :current_staff
